补充readme

+++ WYAKit/Classes/WYAUIKit/WYATextView/WYATextView.h
@@ -35,7 +35,7 @@ NS_ASSUME_NONNULL_BEGIN
 @property (nonatomic, copy)   NSString * title;//标题
 @property (nonatomic, assign) BOOL  showTitle; //default is YES
 @property (nonatomic, assign) BOOL  showWordsCount; // default is YES
-@property (nonatomic, assign) float  textViewMaxHeight;//最大高度
+@property (nonatomic, assign) CGFloat  textViewMaxHeight;//最大高度
 @property (nonatomic, assign) NSUInteger  textViewWordsCount;//最大字数
 @property (nonatomic, readonly) UITextView * textView;
 /**
